requirements.txt
# Dependências principais
fastapi==0.104.1
uvicorn==0.24.0

# Processamento de linguagem natural
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.0.1

# Banco de dados
sqlalchemy>=1.4,<1.5
databases[sqlite]>=0.6.2

# CORS
python-multipart==0.0.6

# Outras utilidades
python-dotenv==1.0.0
pydantic==1.10.7

# NLTK
nltk==3.9.1

# OpenAI para Mistral
openai==1.12.0

